// src/dtos/schedule/itDbResponseData.dto.ts
export class ItDbResponseDto {
  pt_id: string;
  pt_created_at: Date;
  pt_updated_at: Date;
  pt_productType: string;
  pt_categoryId: string;
  id: string;
  productTypeId: string;
  created_at: Date;
  updated_at: Date;
  depYear: number;
  depreciationRate: string;
  nextYearBefore180Days: number;
  nextYearAfter180Days: number;
  curYearAfter180Days: number;
  category_id: string;
  category_created_at: Date;
  category_updated_at: Date;
  category_category_name: string;
}